requirements.txt
djangorestframework
markdown
django-filter
Django
wily